+++ lib/features/profile/presentation/profile_screen.dart
@@ -307,7 +307,7 @@ class _ProfileScreenState extends ConsumerState<ProfileScreen> {
                         _buildModernSettingCard(
                           context: context,
                           title: L10n.of(context)!.version,
-                          subtitle: '1.1.19',
+                          subtitle: '1.1.21',
                           showArrow: false,
                         ),
                         _buildModernSettingCard(